Add files via upload
The file "drip-2.bas" contains the code foe an alternative method of filling the screen with the famous maze characters on the screen. 
Well known under “10 print”. I created two version of the program. One (drip-3-bare.bas) shortened as much as possible (I think). That results also in the most unreadable code.  The second ("drip-2.bas"  a bit more elaborated version. In the second version, with the same result, each part is explained in more detail and therefore also easily be modified to your own liking.